+++ src/System/Text/helper.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace System\Text;
+
+if (! function_exists('string')) {
+
+    function string($text)
+    {
+        return new Text($text);
+    }
+
+    function text($text)
+    {
+        return new Text($text);
+    }
+}